Refactor with modular style
