+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
-import Index from './modules/index/Index';
+import Index from './modules/index';
 
 export default () => (
   <div>
