+++ client/src/components/DateBar.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import leftArrowIcon from '../assets/arrow_left.svg';
-import rightArrowIcon from '../assets/arrow_right.svg';
+import leftArrowIcon from 'assets/arrow_left.svg';
+import rightArrowIcon from 'assets/arrow_right.svg';
 import {
   getQuarterDays,
   getQuarterFirstMonth,
