+++ client/src/modules/core/index.js
@@ -1,0 +1,4 @@
+import NavBar from './components/NavBar';
+import reducer, * as actions from './redux';
+
+export { actions, reducer, NavBar };
