--- client/src/modules/index/QuarterView/styled.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const Grid = styled.div`
-  border-left: solid 1px #f0f3f8;
-  border-radius: 0 0 4px 4px;
-  flex-wrap: wrap;
-  margin: 0;
-  padding: 0;
-  width: 100%;
-`;
