+++ client/src/modules/index/components/EditRole.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import moment from 'moment';
-import { Modal, StateButton } from '../../components';
+import { Modal, StateButton } from 'components';
 import styled from 'styled-components';
 import { Creatable } from 'react-select';
 import 'react-select/dist/react-select.css';
