+++ client/src/modules/index/components/QuarterView/Mobile.js
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import AddToCalendar from 'react-add-to-calendar';
 import './icalstyle.css';
 import moment from 'moment';
-import { findEvent, getCalData } from '../../../utils';
+import { findEvent, getCalData } from 'utils';
 
 export default ({ events, roles, days, onDayClick, onRoleClick }) => {
   const icalicon = { 'calendar-plus-o': 'left' };
