+++ client/src/modules/index/components/QuarterView/icalstyle.css
